pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "office-agent"
version = "0.1.0"
description = "CLI MVP for generating PowerPoint decks from natural-language prompts."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.116,<1",
  "httpx>=0.28,<1",
  "pydantic>=2.12,<3",
  "pywin32>=308,<309; platform_system == 'Windows'",
  "python-multipart>=0.0.20,<1",
  "python-pptx>=1.0.2,<2",
  "requests>=2.32,<3",
  "uvicorn>=0.35,<1",
]

[project.scripts]
office-agent = "office_agent.cli:main"
office-agent-api = "office_agent.api.server:main"

[dependency-groups]
dev = []

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
